Fix package naming inconsistencies and initial build errors
Package Structure Fixes:
- Standardize all test files to use com.example.talktobook package
- Move test files from yama_ai to example directory structure
- Fix 15+ package declaration inconsistencies

Build Configuration:
- Remove duplicate compileOptions and kotlinOptions in build.gradle.kts
- Clean up redundant configuration blocks

Code Fixes:
- Fix Dagger injection in BaseViewModel (protected → internal for crashlyticsManager)
- Remove redundant crashlyticsManager parameters in ViewModels
- Fix nullable reference handling in RecordingViewModel
- Fix string interpolation corruption in Accessibility.kt
- Fix navigation parameter mismatches in TalkToBookNavigation.kt
- Add missing imports in VoiceCorrectionPanel.kt

Test Infrastructure:
- Update test scripts for proper package structure
- Maintain comprehensive test coverage with corrected imports

Documentation:
- Add build error fix report with detailed analysis and solution strategy


(cherry picked from commit c4dca9867eeeafcc9b7b8b68759ef633b4442207)
