+++ app/build.gradle.kts
@@ -42,13 +42,6 @@ android {
     buildFeatures {
         compose = true
     }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = "11"
-    }
     lint {
         disable += "NullSafeMutableLiveData"
     }
