+++ app/src/main/java/com/example/talktobook/data/repository/TranscriptionRepositoryImpl.kt
@@ -11,7 +11,6 @@ import com.example.talktobook.domain.model.Recording
 import com.example.talktobook.domain.model.TranscriptionStatus
 import com.example.talktobook.domain.repository.TranscriptionRepository
 import com.example.talktobook.domain.util.RetryPolicy
-import com.example.talktobook.domain.util.RetryPolicies
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
